Add UserMessages to Administration

+++ Services/BulgarianWines.Services.Data/UserMessagesService.cs
@@ -2,6 +2,7 @@
 {
     using System;
     using System.Collections.Generic;
+    using System.Linq;
     using System.Text;
     using System.Threading.Tasks;
 
@@ -17,23 +18,38 @@
             this.userMessagesRepository = userMessagesRepository;
         }
 
+        public IEnumerable<UserMessage> All() => this.userMessagesRepository.AllWithDeleted();
+
         public async Task Add(UserMessage userMessage)
         {
             await this.userMessagesRepository.AddAsync(userMessage);
             await this.userMessagesRepository.SaveChangesAsync();
         }
 
-        public async Task Delete(UserMessage userMessage)
+        public async Task Delete(string id)
         {
+            var userMessage = this.GetById(id);
+            if (userMessage == null)
+            {
+                return;
+            }
+
             this.userMessagesRepository.Delete(userMessage);
             await this.userMessagesRepository.SaveChangesAsync();
         }
 
         public async Task SetToRead(string id, bool isRead)
         {
-            var userMessage = await this.userMessagesRepository.GetByIdWithDeletedAsync(id);
+            var userMessage = this.GetById(id);
+            if (userMessage == null)
+            {
+                return;
+            }
+
             userMessage.IsRead = isRead;
             await this.userMessagesRepository.SaveChangesAsync();
         }
+
+        public UserMessage GetById(string id) => this.userMessagesRepository.All().FirstOrDefault(x => x.Id == id);
     }
 }
